+++ lib/OrganizationsFieldArray/OrgInfo.js
@@ -1,25 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { get, isEmpty } from 'lodash';
+import { isEmpty } from 'lodash';
 import { AppIcon } from '@folio/stripes/core';
 import { Field } from 'redux-form';
 import {
   Button,
   Card,
   Layout,
-  MultiColumnList,
-  Icon,
 } from '@folio/stripes/components';
 
-export default class InterfacesCard extends React.Component {
+export default class OrgInfo extends React.Component {
   static propTypes = {
     index: PropTypes.string,
     name: PropTypes.string,
     organization: PropTypes.object,
-    organizations: PropTypes.arrayOf(PropTypes.object),
     onReplaceField: PropTypes.func,
     onAddField: PropTypes.func,
+    organizationPickerComponent: PropTypes.elementType,
   };
 
   renderOrganizationPicker = (index, name) => {
@@ -47,52 +45,6 @@ export default class InterfacesCard extends React.Component {
     this.props.onReplaceField(index, { org: {} });
   }
 
-  renderOrgPreview = (organization) => {
-    const { org: { id } } = organization;
-    const { organizations } = this.props;
-    const orgToDisplay = organizations.find(org => org.org.id === id);
-    const interfaces = orgToDisplay ? orgToDisplay.interfaces : [];
-    return (
-      organization.org ?
-        <div>
-          {interfaces && interfaces.length ?
-            <MultiColumnList
-              columnMapping={{
-                name: <FormattedMessage id="ui-agreements.interface.name" />,
-                username: <FormattedMessage id="ui-agreements.interface.username" />,
-                password: <FormattedMessage id="ui-agreements.interface.password" />,
-                type: <FormattedMessage id="ui-agreements.interface.type" />,
-                notes: <FormattedMessage id="ui-agreements.interface.notes" />,
-              }}
-              columnWidths={{
-                name: 130,
-                notes: 180,
-                username: 100,
-                password: 100,
-                type: 130,
-              }}
-              contentData={interfaces}
-              formatter={{
-                name: item => (
-                  <span>
-                    {item.name}
-                    <a href={item.uri}>
-                      <Icon icon="external-link" iconPosition="end" />
-                    </a>
-                  </span>
-                ),
-                notes: item => get(item, 'notes'),
-                username: item => get(item, 'username'),
-                password: item => get(item, 'password'),
-                type: item => get(item, 'type'),
-              }}
-              interactive={false}
-              visibleColumns={['name', 'username', 'password', 'type', 'notes']}
-            /> : <FormattedMessage id="ui-agreements.interface.notFound" />}
-        </div> : null
-    );
-  }
-
   renderOrganizationName = (orgName) => {
     const { index, name } = this.props;
     return (
@@ -108,18 +60,15 @@ export default class InterfacesCard extends React.Component {
     );
   }
 
-  renderInterfacesCard = () => {
+  renderOrgInfo = () => {
     const { index, organization, name } = this.props;
     const { org } = organization;
-    const addOrg = org && !org.orgsUuid_object;
     return (
       <Card
         cardStyle="positive"
         hasMargin
         headerStart={
-          addOrg ?
-            <strong>Organization</strong> :
-            this.renderOrganizationName(org.name)
+          <strong>Organization</strong>
         }
         key={index}
         headerEnd={
@@ -134,7 +83,7 @@ export default class InterfacesCard extends React.Component {
         }
         roundedBorder
       >
-        { addOrg ? this.renderOrganizationName(org.name) : this.renderOrgPreview(organization)}
+        {this.renderOrganizationName(org.name)}
       </Card>
     );
   }
@@ -163,7 +112,7 @@ export default class InterfacesCard extends React.Component {
         headerEnd={this.renderOrganizationPicker(index, name)}
         roundedBorder
       >
-        <React.Fragment>
+        <div>
           <Layout className="textCentered">
             <strong>
               <FormattedMessage id="stripes-erm-components.organizations.noOrganization" />
@@ -172,7 +121,7 @@ export default class InterfacesCard extends React.Component {
           <Layout className="textCentered">
             <FormattedMessage id="stripes-erm-components.organizations.addOrganizationToStart" />
           </Layout>
-        </React.Fragment>
+        </div>
       </Field>
     );
   }
@@ -182,7 +131,7 @@ export default class InterfacesCard extends React.Component {
     const { org } = organization;
 
     return org && !isEmpty(org) ?
-      this.renderInterfacesCard() :
+      this.renderOrgInfo() :
       this.renderDefault();
   }
 }
