+++ lib/OrganizationsFieldArray/tests/OrganizationsFieldArray-test.js
@@ -3,14 +3,13 @@ import { describe, beforeEach, it } from '@bigtest/mocha';
 import { expect } from 'chai';
 import { FieldArray } from 'redux-form';
 import { TextField } from '@folio/stripes/components';
-
 import { mountWithContext } from '../../../tests/helpers';
 import TestForm from '../../../tests/TestForm';
 import OrganizationsFieldArray from '../OrganizationsFieldArray';
 import OrganizationsFieldArrayInteractor from './interactor';
 
-const addOrganizationBtnLabel = 'Add organization';
-const isEmptyMessage = 'No Organizations';
+const addOrganizationBtnLabel = 'Add an organization to get started';
+const isEmptyMessage = 'No Organizations added';
 const items = [
   {
     id: 'ffb9762b-d499-4451-be22-b3cbc6658093',
@@ -41,7 +40,7 @@ const roles = [
   }
 ];
 
-describe('OrganizationsFieldArray', () => {
+describe.only('OrganizationsFieldArray', () => {
   const interactor = new OrganizationsFieldArrayInteractor();
 
   beforeEach(async () => {
