+++ lib/OrganizationsFieldArray/tests/interactor.js
@@ -10,9 +10,9 @@ import {
 } from '@bigtest/interactor';
 
 export default interactor(class OrganizationsFieldArrayInteractor {
-  static defaultScope = '[data-test-organizations-field-array]';
+  static defaultScope = '[data-test-org-fa]';
 
-  hasAddButton = isPresent('[data-test-orgfa-add-button]');
+  hasAddButton = isPresent('[data-test-org-fa-add-button]');
   hasOrg = isPresent('[data-test-org-name]');
   hasRole = isPresent('[data-test-org-role]');
   orgValue = value('[data-test-org-name]');
@@ -21,9 +21,9 @@ export default interactor(class OrganizationsFieldArrayInteractor {
   roleText = text('[data-test-org-role]');
   setOrg = fillable('[data-test-org-name]');
   setRole = selectable('[data-test-org-role]');
-  clickAddButton = clickable('[data-test-orgfa-add-button]');
-  addOrganizationBtnLabel = text('[data-test-orgfa-add-button]');
+  clickAddButton = clickable('[data-test-org-fa-add-button]');
+  addOrganizationBtnLabel = text('[data-test-org-fa-add-button]');
   emptyMessage = text('[data-test-org-empty-message]');
-  count = count('[data-test-org]');
+  count = count('[data-test-organizations-org]');
   clickDeleteButton = clickable('[data-test-org-delete-button]');
 });
