+++ package.json
@@ -43,7 +43,6 @@
     "prop-types": "^15.6.0",
     "react-dropzone": "^10.0.0",
     "react-intl": "^2.4.0",
-    "react-router-dom": "^5.0.1",
     "redux-form": "^7.0.3"
   },
   "peerDependencies": {
